# swagger.json
{
	"openapi": "3.0.3",
	"info": {
	  "title": "Slim Mom`s - OpenAPI 3.0",
	  "version": "1.0.0",
	  "description": "This is a description of the InSight team's Slim Mom's project server, based on the OpenAPI 3.0 specification.\n\nThis server allows you to get information about the daily calorie intake and a list of foods to avoid based on physical data and blood type.\n\nAlso, the ability to save, receive and edit the list of foods eaten for a specific day is implemented.\nThe information is tied to a specific user (a token is used).\n\nImplemented registration, user login/logout\n\nSome useful links:\n- [The Slim Mom's repository](https://github.com/AegisVP/InSight-web)\n- [The source API definition for the Slim Mom's](https://github.com/AegisVP/InSight-api/blob/main/swagger.json)"
	},
	"servers": [
	  {
		"url": "https://insight-txsh.onrender.com"
	  }
	],
...
	"paths": {
...
	  "/user/login": {
		"post": {
		  "tags": [
			"user"
		  ],
		  "summary": "Logs user into the system",
		  "description": "Logs user into the system",
		  "operationId": "loginUser",
		  "requestBody": {
			"required": true,
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/UserLoginRequest"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "successful operation",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/UserAuthResponse"
				  }
				}
			  }
			},
			"400": {
			  "description": "Invalid email/password supplied"
			},
			"404": {
			  "description": "User not found"
			},
			"500": {
			  "description": "Server error"
			}
		  }
		}
	  },
...
	},
	"components": {
	  "schemas": {
...
		"UserLoginRequest": {
		  "type": "object",
		  "properties": {
			"email": {
			  "type": "string",
			  "example": "example@email.com"
			},
			"password": {
			  "type": "string",
			  "example": "12345"
			}
		  },
		  "xml": {
			"name": "userLoginRequest"
		  }
		},
		"UserAuthResponse": {
		  "type": "object",
		  "properties": {
			"email": {
			  "type": "string",
			  "example": "example@email.com"
			},
			"name": {
			  "type": "string",
			  "example": "Alex"
			},
			"token": {
			  "type": "string",
			  "example": "12345"
			}
		  },
		  "xml": {
			"name": "UserAuthResponse"
		  }
		},
...
	  }
	}
  }